refactor(arcgcode): renamed new pipeline to CuraPostProcessor
